Moved performing the HTTP requests inside the try..catch block (fixes #7)
In the HttpServiceClient, the initial request was performed outside the try..catch block. This leads to leaky abstractions when an error occurs. We want to throw a generic HaveIBeenPwnedClientException in that case but we can't since it was outside of the try..catch. This change moves performing the request to inside the try.