Removed disposing of an injected HttpClient (fixes #6)
When the HttpServiceClient was disposed, it also disposed the injected HttpClient. That's rude and error-prone, since the callee might use it outside of the HttpServiceClient.

In this change, we added a constructor that creates the HttpClient and there's still an overloaded constructor that  accepts an injected HttpClient. The HttpClient created by the HttpServiceClient itself will be disposed when the HttpServiceClient is disposed, but any injected HttpClient instances will not.