Remove IDisposable from IHaveIBeenPwnedClient (fixes #10)
Previously, IHaveIBeenPwnedClient extended from IDisposable. That was undesirable, because it forces any implementation of a custom client to implement the dispose logic too and some might not need it. We could safely remove the interface (it's breaking, so major version bump), and in HaveIBeenPwnedClient.Dispose(bool) only dispose the used IHaveBeenPwnedClient when it's disposable.