Remove ClientSettings from HaveIBeenPwnedClient (Fixes #12)
The constructor of HaveIBeenPwnedClient only accepted a ClientSettings object because it needs to pass it on to HttpServiceClient. We removed it from HaveIBeenPwnedClient and if a user wants to override the defaults they can instantiate HttpServiceClient themselves and use constructor injection to pass that client to HaveIBeenWPnedClient.

We also moved ClientSettings to the AtleX.HaveIBeenPwned.Communication.Http namespace and renamed it to HttpClientSettings.