Reorganise classes and namespaces (fixes #16)
Previously, a user had to add a lot of usings when using this lib. For example, the main HaveIBeenPwnedClient was in AtleX.HaveIBeenPwned, the resulting Breach and Paste objects were in AtleX.HaveIBeenPwned.Data (granted, use var), the exceptions were somewhere in AtleX.HaveIBeenPwned.Communiction, etc. This was confusing and unnecessary so I've moved those classes to their appropriate namespaces.

Related to this, I renamed the AtleX.HaveIBeenPwned.Communication namespace to AtleX.HaveIBeenPwned.Clients. It contains the clients, the name should reflect that.