Marked KAnonimityHelper as internal and move it to the AtleX.HaveIBeenPwned.Clients.Http namespace (fixes #18)
The KAnonimityHelper was public but it's quite tightly coupled to the HttpHaveIBeenPwnedClient. It was most likely of limited use outside of it, since any custom IHaveIBeenPwnedClient will probably work differently otherwise the user would have used the HttpHaveIBeenPwnedClient. So, we reduced the API surface of the lib by making internal stuff internal.