Fixed GetBreachesInternalAsync(string, BreachMode, CancellationToken) return null instead of an empty enumerable when no data was loaded (fixes #23)
We returned null when no data was loaded instead of an empty enumerable. That lead to bugs on the call site because people are expecting to be able to use LINQ or something like that to search in the returned collection. Also, GetPastesInternalAsync(string, CancellationToken) does correctly return an empty enumerable so we were inconsistent too.