Fold HaveIBeenPwnedClient and HttpHaveIBeenPwnedClient into one (fixes #26)
Originally, I created the HaveIBeenPwnedClient as a facade/factory class for when multiple IHaveIBeenPwnedClient implementations would be added. But realistically, we will only have the HTTP implementation in the foreseeable future unless HaveIBeenPwned.com significantly changes the way it works and we can't predict that so whatever. It was just my fault of overengineering this.

Previously, having HaveIBeenPwnedClient just passing all calls to HttpHaveIBeenPwnedClient was becoming more and more of a burden. Using it in .NET Core's DI system wass hard because that will always call the most specific constructor (the one with the IHaveIBeenPwnedClient). So in the end you just registered HttpHaveIBeenPwnedClient for DI and not bothered with HaveIBeenPwnedClient at all. Also, we had to do a lot of unit testing etc. for something that didn't serve any real purpose.

So, we replaced the business logic of HaveIBeenPwnedClient with HttpHaveIBeenPwnedClient.