Increased performance of HaveIBeenPwnedClient.IsPwnedPasswordAsync(string. [CancellationToken])
Previously, we retrieved a Stream from the HaveIBeenPwned.com API and opened a StreamReader to read it. We read each line async, compared it to the KAnonimity prefix we calculated and either read the next line async or exited when we had a match. However, this meant constructing a state machine for each line and an expensive string compare and on a average response this took 113 us. and consumed 142 KB of memory on .NET Core 3.0.

This change reads the whole response and searches the resulting string for the full hash.  This creates only two async state machines (one for GetAsync(RequestMessage, CancellationToken), one for StreamReader.ReadToEndAsync()) and then searches in the string. The resulting string can be huge, but searching it is especially effective on .NET Core.

On .NET Core 3.0, the time needed cuts back from around 113 us. on the test machine to 60 us. The allocated memory is going down from 142 KB to 137. Moreover, on .NET 4.7.2 the time needed is only 15 us. compared to 313 us. in the old version and allocations are down to 146 KB from 161 KB.