Reduce allocations
Previously, we read the whole response of a call to the HaveIBeenPwned.com API in HaveIBeenPwnedClient to a new MemoryStream. This is wasteful. since we copy a full response to a new MemoryStream.

This change just returns the existing Stream instead of copying the response. We parse it where we need it, resulting in around a around 20% reduction in allocations and a 5-60% reduction in time needed, depending on the method called and the target framework.