Removed the default value from HaveIBeenPwnedClientSettings.ApplicationName and throw an exception when it's null or empty (fixes #30)
Previously,, HaveIBeenPwnedClientSettings.ApplicationName defaulted to "AtleX.HaveIBeenPwned" so it was not mandatory for users to set it. However, the FAQ of the HaveIBeenPwned.com API makes it absolutely clear that a user should set this:

"Not properly identifying the user agent such that it accurately describes the consumer of the API"

This removes  the default application name and throws an exception from HaveIBeenPwnedClient when a null or empty application name is set.