No longer clear the default  request headers on the HttpClient (fixes #35)
A HttpClient can be injected and it might already be fully configured. By clearing all request headers we remove the configuration done by the user. E.g., support for compression is removed.

This change removes clearing the default request headers. Instead the User-Agent header gets overwritten by the application name from the HaveIBeenPwnedClientSettings and the application/json media type is added to the accepted media types.