Override ToString() on Breach and Paste to display the Name or Title (fixes #41)
(Inspired by #40)

When calling ToString() on Breach or Paste, the type name was returned because these methods weren't overridden.  This might cause confusion, because in the debugger a value is visible (because of DebuggerDisplayAttribute usage) but e.g. using the object in string interpolation will not display that same value.

This change overrides ToString() on Breach to return the value of the Name property. On Paste, ToString() returns the value of the Title property. This is the same as their respective DebuggerDisplayAttribute values.