Fixed not creating a UriBuilder when it's not needed
Previously, in HaveIBeenPwnedClient.GetBreachesInternalAsync(string, BreachMode, CancellationToken), a new UriBuilder is always created. But we only needed that UriBuilder when the BreachMode was BreachMode.ExcludeUnverified or (currently) BreachMode.All. Otherwise just a Uri is enough. A UriBuilder is quite expensive in both time and memory so removing it from the default path is nice.

This change only creates a new UriBuilder when it is needed. Otherwise it will just use a Uri. For the common case, this improves performance on .NET Core 3.0 by around 30% and drops memory usage by around 30%.