Removed no longer applicable override of the Message property
Previously, the message in RateLimitException was created in the constructor. However, that was later optimized to an override of the Message property since the message is not used in 99% of the cases. The exception will be handled in code and something will be retried after the RetryAfter timespan is elapsed.

Later on, I realized that including the timespan in the message serves no purpose at all so I changed it to a hardcoded message in e61f90d08ef39c8a98e55a83938e4f389f983404.

However, that also meant that overriding the Message property serves no purpose anymore. Instead we can just pass the message to the constructor of the base class. So, that's what we do now.