Re-use JsonSerializerOptions
Previously, the JsonSerializerOptions were created per instance of HaveIBeenPwnedClient. Since the (de)serializtion info is stored in the options this caused them to be regenerated for each instance of HaveIBeenPwnedClient.

Now we store a single JsonSerializerOptions for the lifetime of the application in a static readonly field. Also, the context was expanded to cover IEnumerable<T> of the different types to prevent errors on .NET 7.0 where an exception is thrown for (de)serialization types not explicitly configiured in the context.