Use custom JsonConvertor for domain users (fixes #94)
Previously, breached users of a domain were first deserialized to a dictonary with string keys for the username and IEnumerables for the breached domains of that user. Then this dictionary was converted into an IEnumerable of DomainUser. This two-stage approach was a bit slower than necessary and also consumed more resources.

In this commit, a custom JsonConvertor is introduced that directly converts the JSON received from the HaveIBeenPwned API to an IEnumerable of DomainUser.