Reworked the benchmarks to allow running them for each breach/leak type individually (fixes #95)
Previously, there's was a single IHaveIBeenPwnedClientBenchmarks class that contained all benchmarks for breaches, breached users, pwned passwords, etc. When making changes to the methods related to one of them, we needed to run all the benchmarks to verify performance, which is quite a time consuming process.

This is now changed to separate benchmarks for each !HaveIBeenPwned*Client interface. That allows us to run the benchmarks for a specific breach/leak only if we want to.