Use Convert.ToHextString() and spans to optimize the KAnonimityHelper (fixes #96)
Previously, a StringBuilder with a format was used to build both parts of the KAnonimity hash. However, in .NET 5 a Convert.ToHexString() helper was introduced that is cheaper to use than creating a StringBuilder and iterating over the byte array with the raw hash ourselves. We also use a span to create both parts of the hash.