Use a rented array to store the generated hash (Fixed #99)
By using a rented buffer to store the hash in, instead of creating a new byte array for it, we can reduce allocations even further after the improvements made in #98.

Additionally, renting arrays is made more resilient by wrapping the code in a try..finally and returning all rented arrays in the finally block.